requirements.txt
requests>=2.19.1
psycopg2-binary>=2.7.4
flask>=1.0.2
